backend/api/pyproject.toml
[tool.poetry]
name = "api-server"
version = "0.1.0"
description = "Prototype FastAPI server for Steam Indie game Recommendation System"
authors = ["Sungsu Bae <ssb6720@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "api_server"}]

[tool.poetry.dependencies]
python = ">=3.8,<3.10"
fastapi = "^0.89.1"
requests = "^2.28.2"
uvicorn = "^0.20.0"
jinja2 = "^3.1.2"
python-multipart = "^0.0.5"
pyyaml = "^6.0"
sqlalchemy = "^1.4.46"
alembic = "^1.9.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
black = "^22.12.0"
jupyter = "^1.0.0"
notebook = "^6.5.2"
numpy = "^1.24.1"
pandas = "^1.5.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
